Restructures migrate storage code
